+++ standard/packages/react/tests/index.tsx
@@ -642,7 +642,7 @@ it('eventContent render can return true to use default rendering', () => {
 
 
 ;[
-  'viewContent', // https://github.com/fullcalendar/fullcalendar/issues/7160
+  'content', // https://github.com/fullcalendar/fullcalendar/issues/7160
   'component', // https://github.com/fullcalendar/fullcalendar/issues/7207
 ].forEach((settingName) => {
   it(`can render custom content in a custom view (with ${settingName} setting)`, () => {
@@ -662,7 +662,7 @@ it('eventContent render can return true to use default rendering', () => {
 })
 
 ;[
-  'viewContent',
+  'content',
   'component',
 ].forEach((settingName) => {
   it(`can render custom content AS FUNCTION in a custom view (with ${settingName} setting)`, () => {
@@ -700,7 +700,7 @@ it('custom view receives enough props for slicing', () => {
       ]}
       views={{
         customView: {
-          viewContent: (props) => {
+          content: (props) => {
             const segs = sliceEvents(props, true); // allDay=true
             return (
               <>